tinystdio: Case width to (size_t) when comparing with size in vfprintf
'width' is always non-negative at this point, so it is safe to cast it
to an unsigned value to avoid a compiler warning.

+++ newlib/libc/tinystdio/vfprintf.c
@@ -691,7 +691,7 @@ int vfprintf (FILE * stream, const char *fmt, va_list ap)
 
             str_lpad:
                 if (!(flags & FL_LPAD)) {
-                    while (width > size) {
+                    while ((size_t) width > size) {
                         my_putc (' ', stream);
                         width--;
                     }
